refactor : LottoNumberParser view package로 이동
