+++ src/main/kotlin/next/step/lotto/view/LottoNumberParser.kt
@@ -1,7 +1,9 @@
-package next.step.lotto.domain
+package next.step.lotto.view
+
+import next.step.lotto.domain.LottoNumber
 
 object LottoNumberParser {
-    
+
     fun parse(str: String): Set<LottoNumber> = str.split(",")
         .map { LottoNumber.of(it.trim().toInt()) }
         .toSet()
