+++ src/test/kotlin/next/step/lotto/view/LottoNumberParserTest.kt
@@ -1,4 +1,4 @@
-package next.step.lotto.domain
+package next.step.lotto.view
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.DescribeSpec
